# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EDCS API",
        "description": "The Elite Dangerous Companion Suite API provides galaxy data from Elite Dangerous to various consumers, including the EDCS frontend at https://edcs.app. Data is sourced from EDSM (Elite Dangerous Star Map), EDDN (Elite Dangerous Data Network), and the Frontier Companion API (CAPI).",
        "contact": {
            "name": "EDCS",
            "url": "https://edcs.app"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.edcs.app/api",
            "description": "EDCS API Server"
        }
    ],
    "paths": {
...
        "/bodies/{slug}": {
            "get": {
                "tags": [
                    "Bodies"
                ],
                "summary": "Get a single celestial body by slug",
                "description": "Retrieves a star, planet, or moon by its slug ({body_id}-{name}). Pass withSystem to embed the parent system, withStations to also embed the system's stations.",
                "operationId": "4a99b4c6d64163da0c3817fb51a3d0d4",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Body slug in format {body_id}-{name}",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "108086401534265707-earth"
                        }
                    },
                    {
                        "name": "withSystem",
                        "in": "query",
                        "description": "Embed the parent system",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    },
                    {
                        "name": "withStations",
                        "in": "query",
                        "description": "Embed the parent system's stations",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Celestial body",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/SystemBody"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Body not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Station": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Daedalus"
                    },
                    "type": {
                        "type": "string",
                        "example": "Orbis Starport"
                    },
                    "body": {
                        "type": "string",
                        "example": "Sol",
                        "nullable": true
                    },
                    "distance_to_arrival": {
                        "type": "number",
                        "format": "float",
                        "example": 508
                    },
                    "controlling_faction": {
                        "type": "string",
                        "example": "Sol Constitution Party",
                        "nullable": true
                    },
                    "allegiance": {
                        "type": "string",
                        "example": "Federation",
                        "nullable": true
                    },
                    "government": {
                        "type": "string",
                        "example": "Democracy",
                        "nullable": true
                    },
                    "economy": {
                        "type": "string",
                        "example": "Industrial",
                        "nullable": true
                    },
                    "second_economy": {
                        "type": "string",
                        "example": "Refinery",
                        "nullable": true
                    },
                    "has_market": {
                        "type": "boolean",
                        "example": true
                    },
                    "has_shipyard": {
                        "type": "boolean",
                        "example": true
                    },
                    "has_outfitting": {
                        "type": "boolean",
                        "example": true
                    },
                    "other_services": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Restock",
                            "Repair",
                            "Contacts"
                        ]
                    },
                    "last_updated": {
                        "properties": {
                            "information": {
                                "type": "string",
                                "format": "date-time",
                                "nullable": true
                            },
                            "market": {
                                "type": "string",
                                "format": "date-time",
                                "nullable": true
                            },
                            "shipyard": {
                                "type": "string",
                                "format": "date-time",
                                "nullable": true
                            },
                            "outfitting": {
                                "type": "string",
                                "format": "date-time",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "slug": {
                        "type": "string",
                        "example": "128016384-daedalus"
                    },
                    "system": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/System"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "SystemBody": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "body_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "name": {
                        "type": "string",
                        "example": "Earth"
                    },
                    "type": {
                        "type": "string",
                        "example": "Planet"
                    },
                    "sub_type": {
                        "type": "string",
                        "example": "Earth-like world"
                    },
                    "distance_to_arrival": {
                        "type": "number",
                        "format": "float",
                        "example": 500
                    },
                    "is_main_star": {
                        "type": "boolean",
                        "example": false
                    },
                    "is_scoopable": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "spectral_class": {
                        "type": "string",
                        "example": "G",
                        "nullable": true
                    },
                    "luminosity": {
                        "type": "string",
                        "example": "Va",
                        "nullable": true
                    },
                    "solar_masses": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "solar_radius": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "absolute_magnitude": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "discovery": {
                        "properties": {
                            "commander": {
                                "type": "string",
                                "example": "CMDR Explorer",
                                "nullable": true
                            },
                            "date": {
                                "type": "string",
                                "example": "3302-06-01",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "radius": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "gravity": {
                        "type": "number",
                        "format": "float",
                        "example": 1,
                        "nullable": true
                    },
                    "earth_masses": {
                        "type": "number",
                        "format": "float",
                        "example": 1,
                        "nullable": true
                    },
                    "surface_temp": {
                        "type": "number",
                        "format": "float",
                        "example": 288,
                        "nullable": true
                    },
                    "is_landable": {
                        "type": "boolean",
                        "example": false
                    },
                    "atmosphere_type": {
                        "type": "string",
                        "example": "Suitable for water-based life",
                        "nullable": true
                    },
                    "volcanism_type": {
                        "type": "string",
                        "nullable": true
                    },
                    "terraforming_state": {
                        "type": "string",
                        "example": "Already terraformed",
                        "nullable": true
                    },
                    "axial": {
                        "properties": {
                            "axial_tilt": {
                                "type": "number",
                                "format": "float",
                                "nullable": true
                            },
                            "semi_major_axis": {
                                "type": "number",
                                "format": "float",
                                "nullable": true
                            },
                            "rotational_period": {
                                "type": "number",
                                "format": "float",
                                "nullable": true
                            },
                            "is_tidally_locked": {
                                "type": "boolean",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "orbital": {
                        "properties": {
                            "orbital_period": {
                                "type": "number",
                                "format": "float",
                                "nullable": true
                            },
                            "orbital_eccentricity": {
                                "type": "number",
                                "format": "float",
                                "nullable": true
                            },
                            "orbital_inclination": {
                                "type": "number",
                                "format": "float",
                                "nullable": true
                            },
                            "arg_of_periapsis": {
                                "type": "number",
                                "format": "float",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "rings": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "parents": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "slug": {
                        "type": "string",
                        "example": "108086401534265707-earth"
                    },
                    "system": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/System"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "SystemInformation": {
                "properties": {
                    "allegiance": {
                        "type": "string",
                        "example": "Federation"
                    },
                    "government": {
                        "type": "string",
                        "example": "Democracy"
                    },
                    "population": {
                        "type": "integer",
                        "example": 22780000000
                    },
                    "security": {
                        "type": "string",
                        "example": "High"
                    },
                    "economy": {
                        "type": "string",
                        "example": "Refinery"
                    },
                    "controlling_faction": {
                        "properties": {
                            "name": {
                                "type": "string",
                                "example": "Mother Gaia"
                            },
                            "state": {
                                "type": "string",
                                "example": "None"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "SystemCoords": {
                "properties": {
                    "x": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    },
                    "y": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    },
                    "z": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    }
                },
                "type": "object"
            },
            "System": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "id64": {
                        "type": "integer",
                        "example": 10477373803
                    },
                    "name": {
                        "type": "string",
                        "example": "Sol"
                    },
                    "coords": {
                        "$ref": "#/components/schemas/SystemCoords"
                    },
                    "slug": {
                        "type": "string",
                        "example": "10477373803-sol"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "information": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/SystemInformation"
                            }
                        ],
                        "nullable": true
                    },
                    "bodies": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SystemBody"
                        },
                        "nullable": true
                    },
                    "stations": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Station"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum bearer token. Obtain via POST /auth/login or the Frontier SSO flow.",
                "scheme": "bearer",
                "bearerFormat": "Token"
            }
        }
    },
...
}